mmaintegration: move clamped mult model to legacy
